requirements.txt
autopep8==1.4.3
certifi==2019.3.9
Click==7.0
colorama==0.4.1
cssmin==0.2.0
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-JWT==0.3.2
itsdangerous==1.1.0
Jinja2==2.10
jsmin==2.2.2
MarkupSafe==1.1.0
peewee==3.8.2
peewee-db-evolve==3.7.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.4.2
python-dotenv==0.10.1
six==1.12.0
webassets==0.12.1
Werkzeug==0.14.1
